fixup! Allow overriding fragment from input via TypeOverrides.CacheIdentifier
