+++ integration-tests/type-tests/fromOptionValue/override/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true
+  },
+  "references": [
+    {
+      "path": "../../tsconfig.src.json"
+    }
+  ]
+}
