+++ src/core/index.ts
@@ -199,12 +199,14 @@ import type {
   PossibleDefaultOptions,
 } from "./defaultOptions.js";
 import type { QueryManager } from "./QueryManager.js";
+import type { OverridableTypes } from "./types.js";
 import type { NextFetchPolicyContext } from "./watchQueryOptions.js";
 /** @internal */
 export declare namespace InternalTypes {
   export type {
     DefaultOptionsParentObject,
     NextFetchPolicyContext,
+    OverridableTypes,
     PossibleDefaultOptions,
     QueryManager,
   };
